requirements.txt
djangorestframework==3.12.4
django-filter==21.1
Markdown==3.3.6
django-cors-headers==3.10.0
django-extensions==3.1.5
pandas==0.25.3
nltk==3.6.5
gensim==4.1.2